--- md407/util.c
@@ -1,6 +1,0 @@
-#include "util.h"
-
-void test() {
-	char c = 'A';
-	c++;
-}
